Merge week 7 progress to main

See merge request cse1105/2021-2022/team-repositories/oopp-group-51/repository-template!75

+++ commons/src/main/java/commons/Evaluation.java
@@ -11,16 +11,16 @@ import static org.apache.commons.lang3.builder.ToStringStyle.MULTI_LINE_STYLE;
 
 public class Evaluation {
     public int points;
-    public List<Integer> correctAnswers;
+    public List<Long> correctAnswers;
     public Question.QuestionType type;
 
     public Evaluation() {
         this.type = Question.QuestionType.UNKNOWN;
-        this.correctAnswers = new ArrayList<Integer>();
+        this.correctAnswers = new ArrayList<Long>();
         this.points = 0;
     }
 
-    public Evaluation(int points, Question.QuestionType type, List<Integer> correctAnswers) {
+    public Evaluation(int points, Question.QuestionType type, List<Long> correctAnswers) {
         this.type = type;
         this.correctAnswers = correctAnswers;
         this.points = points;
@@ -28,6 +28,7 @@ public class Evaluation {
 
     /**
      * Equals method
+     *
      * @param obj - Object that will be compared with this
      * @return true if this and obj are equal
      */
@@ -38,6 +39,7 @@ public class Evaluation {
 
     /**
      * Hashcode method
+     *
      * @return the hashcode of the activity
      */
     @Override
@@ -47,6 +49,7 @@ public class Evaluation {
 
     /**
      * ToString method
+     *
      * @return String containing the activity in a readable format
      */
     @Override
